requirements.txt
PyQt5
rembg
Pillow
